Trim and condense.
