+++ Backend.Tests/Controllers/LiftControllerTests.cs
@@ -23,7 +23,6 @@ namespace Backend.Tests.Controllers
         private IWordRepository _wordRepo = null!;
         private IWordService _wordService = null!;
         private IProjectRepository _projRepo = null!;
-        private IProjectService _projService = null!;
         private ILiftService _liftService = null!;
         private LiftController _liftController = null!;
         private IHubContext<CombineHub> _notifyService = null!;
@@ -38,14 +37,13 @@ namespace Backend.Tests.Controllers
         {
             _permissionService = new PermissionServiceMock();
             _projRepo = new ProjectRepositoryMock();
-            _projService = new ProjectServiceMock();
             _projId = _projRepo.Create(new Project { Name = _projName }).Result!.Id;
             _wordRepo = new WordRepositoryMock();
             _liftService = new LiftService();
             _notifyService = new HubContextMock();
             _logger = new MockLogger();
             _liftController = new LiftController(
-                _wordRepo, _projRepo, _projService, _permissionService, _liftService, _notifyService, _logger);
+                _wordRepo, _projRepo, _permissionService, _liftService, _notifyService, _logger);
             _wordService = new WordService(_wordRepo);
         }
