--- Backend/Interfaces/IProjectService.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace BackendFramework.Interfaces
-{
-    public interface IProjectService
-    {
-        Task<bool> CanImportLift(string projectId);
-    }
-}
